Add a build option to use blake2 prehashing for ed25519 signatures
The default is sha512 but the ed25519 specification allows
alternative hash algorithms to be used, and some crypto applications
are now using blake2.

Ideally this would be a runtime choice, but that has a performance
impact and it's rare that a library client would need to use multiple
types of hashing with ed25519.