instance (MonadRandom m, MonadTrans t, Monad (t m)) => MonadRandom (t m)
It is common to use monad transformer stacks in applications, with
IO at the heart.  For applications using cryptonite, it is useful to
have MonadRandom instances for the common transformers.  Add the
instance

  instance (MonadRandom m, MonadTrans t, Monad (t m)) => MonadRandom (t m)

which achieves this with little code and one small additional
dependency, 'transformers', with no new transitive dependencies and
upon which many programs are likely to end up depending anyway.
Although new dependencies are discouraged, this change is motivated
by the usefulness of the instance and the desire to avoid orphan
instances in applications.

Also drive-by remove (Functor m) from MonadRandom class constraints,
because Functor is now a superclass of Monad and cryptonite no
longer supports pre-AMP GHC versions.

Fixes: https://github.com/haskell-crypto/cryptonite/issues/304